# Speed of CommandsCache increased in case when aliases have multiple updates (i.e. conda init).