--- a/news/commands_cache_speed_up.rst
+++ b/news/commands_cache_speed_up.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* <news item>
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* Speed of CommandsCache increased when aliases have multiple updates (i.e. conda init).
+
+**Security:**
+
+* <news item>