--- a/news/commands_cache_speed_up.rst
+++ b/news/commands_cache_speed_up.rst
@@ -16,7 +16,7 @@
 
 **Fixed:**
 
-* Speed of CommandsCache increased when aliases have multiple updates (i.e. init conda).
+* Speed of CommandsCache increased when aliases have multiple updates (i.e. conda init).
 
 **Security:**