--- a/xonsh/commands_cache.py
+++ b/xonsh/commands_cache.py
@@ -75,6 +75,10 @@ class CommandsCache(cabc.Mapping):
                 ret.append(e)
         return ret
 
+    @staticmethod
+    def get_cmd(cmd):
+        return cmd.upper() if ON_WINDOWS else cmd
+
     @property
     def all_commands(self):
         paths = builtins.__xonsh__.env.get("PATH", [])
@@ -82,12 +86,12 @@ class CommandsCache(cabc.Mapping):
         path_immut = tuple(x for x in paths if os.path.isdir(x))
         # did PATH change?
         path_hash = hash(path_immut)
-        cache_valid = path_hash == self._path_checksum
+        cache_valid_path = path_hash == self._path_checksum
         self._path_checksum = path_hash
         # did aliases change?
         alss = getattr(builtins, "aliases", dict())
         al_hash = hash(frozenset(alss))
-        cache_valid = cache_valid and al_hash == self._alias_checksum
+        cache_valid_aliases = al_hash == self._alias_checksum
         self._alias_checksum = al_hash
         # did the contents of any directory in PATH change?
         max_mtime = 0
@@ -95,21 +99,34 @@ class CommandsCache(cabc.Mapping):
             mtime = os.stat(path).st_mtime
             if mtime > max_mtime:
                 max_mtime = mtime
-        cache_valid = cache_valid and (max_mtime <= self._path_mtime)
+        cache_valid_paths = max_mtime <= self._path_mtime
         self._path_mtime = max_mtime
-        if cache_valid:
-            return self._cmds_cache
+
+        if cache_valid_path and cache_valid_paths:
+            if cache_valid_aliases:
+                return self._cmds_cache
+            else:
+                for cmd, alias in alss.items():
+                    cmd = self.get_cmd(cmd)
+                    if cmd in self._cmds_cache:
+                        self._cmds_cache[cmd] = (self._cmds_cache[cmd][0], alias)
+                    else:
+                        self._cmds_cache[cmd] = (cmd, True)
+                return self._cmds_cache
+
         allcmds = {}
         for path in reversed(path_immut):
             # iterate backwards so that entries at the front of PATH overwrite
             # entries at the back.
             for cmd in executables_in(path):
-                key = cmd.upper() if ON_WINDOWS else cmd
+                key = self.get_cmd(cmd)
                 allcmds[key] = (os.path.join(path, cmd), alss.get(key, None))
+
         for cmd in alss:
             if cmd not in allcmds:
-                key = cmd.upper() if ON_WINDOWS else cmd
-                allcmds[key] = (cmd, True)
+                key = self.get_cmd(cmd)
+                allcmds[key] = (self.get_cmd(cmd), True)
+
         self._cmds_cache = allcmds
         return allcmds