Prerequisites:
* mostly #3912
* a bit of #3915

Before this PR (values after 10 runs):
```
$ xonsh --timings --rc ~/.xonshrc_conda 
*
*
*
|------------------|-----------|-----------|
|Event name        | Time (s)  | Delta (s) |
|------------------|-----------|-----------|
|start             |   0.000   |   0.000   |
|pre_execer_init   |   0.015   |   0.015   |
|post_execer_init  |   0.057   |   0.042   |
|on_pre_rc         |   0.057   |   0.000   |
|on_post_rc        |   0.253   |   0.196   |
|on_ptk_create     |   0.334   |   0.081   |
|on_post_init      |   0.334   |   0.000   |
|on_pre_cmdloop    |   0.334   |   0.000   |
|on_pre_prompt     |   0.738   |   0.403   |
|------------------|-----------|-----------|
```

After this PR (values after 10 runs):
```
$ xonsh --timings --rc ~/.xonshrc_conda 
*
|------------------|-----------|-----------|
|Event name        | Time (s)  | Delta (s) |
|------------------|-----------|-----------|
|start             |   0.000   |   0.000   |
|pre_execer_init   |   0.014   |   0.014   |
|post_execer_init  |   0.057   |   0.043   |
|on_pre_rc         |   0.057   |   0.000   |
|on_post_rc        |   0.223   |   0.166   |
|on_ptk_create     |   0.306   |   0.083   |
|on_post_init      |   0.306   |   0.000   |
|on_pre_cmdloop    |   0.306   |   0.000   |
|on_pre_prompt     |   0.389   |   0.082   |
|------------------|-----------|-----------|
```

As result: 0.389 vs 0.738 = 50% speed up on Linux.

Note! To see this speed up you should use #3916.

Closes #3912