@scopatz 2020-10-21
Hey @anki-code - thanks for putting in this PR.  The code looks good to me, but I was having a hard time reproducing.  For me, 

```
scopatz@artemis:~$ xonsh --timings
 Debug level: Off
|------------------|-----------|-----------|
|Event name        | Time (s)  | Delta (s) |
|------------------|-----------|-----------|
|start             |   0.000   |   0.000   |
|pre_execer_init   |   0.011   |   0.011   |
|post_execer_init  |   0.046   |   0.035   |
|on_pre_rc         |   0.046   |   0.000   |
|on_post_rc        |   0.195   |   0.149   |
|on_post_init      |   0.204   |   0.009   |
|on_pre_cmdloop    |   0.204   |   0.000   |
|on_pre_prompt     |   0.204   |   0.000   |
|------------------|-----------|-----------|
```

So the `on_pre_prompt` diff is `0.000`. I have conda in my xonshrc.  Are you only seeing this issue on Windows WSL?