@anki-code 2020-10-21
Probably I'm not described the case step by step and with details. I try again.

1. I've found the issue in WSL and we added the warning about huge amount of files in pr #3911
2. Then I backed to pure Linux Ubuntu and found that CommandsCache (CC) works slow for me because it has two issues.
3. The first issue is that CC read the directory from PATH and the symlink to thisdirectory (double work). I've fixed it in pr #3915
4. The second issue is that CC invalidates cache and read files from all PATH directories every time when aliases have changed. I've caught this behavior on sourcing conda. The files from PATHs was scanned 3 times because conda adds the alias and invalidates the cache. This PR fix this. If there is no changes in PATH and PATH directories it will not scanned. As result I have x2 speed up the startup of the shell on Linux.

The case completely described in #3912.