@anki-code 2020-10-21
And finally I've found that this speed issue was not shown in `--timings` and I made the pr #3916 to fix it too.