@anki-code 2020-10-21
How prompt relates with CommandsCache (CC)? Prompt renders tokens i.e. git branch. To do this it calls the git command. The git command calls CC. And at the end of stack CC read the directories and files from PATH many times and slow because in linux there are around 3000 executable files in the PATH directories. I fixed this all.