@scopatz 2020-10-22
Thanks for the explanation! Just to be clear, you only saw the slow down on WSL though? I am in favor of the change, I just want to make sure