@anki-code 2020-10-22
The tables with timing in the first message of this thread are from pure Ubuntu:
* 0.738s - before this PR (reading of PATH directories 3 times)
* 0.389s - after this PR (reading of PATH directories 1 time)