@anki-code 2020-10-22
>  I was having a hard time reproducing. For me...

This is wrong time. If you want to test your xonsh timing:
1. Merge #3916 and post new table
2. Run [the code from 3915](https://github.com/xonsh/xonsh/issues/3915#issue-726199438) to get the amount of files in your PATH directories.
3. Install xonsh from master and run `COMMANDS_CACHE_SIZE_WARNING=1 xonsh --timings` and post here the count of warnings you see.