@anki-code 2020-10-22
Linux test

Before PR - 0.570:

<details>

```
(base) pc@pc:~$ COMMANDS_CACHE_SIZE_WARNING=1 xonsh --timings --rc ~/.xonshrc_conda 
Warning! Found 2,935 executable files in the PATH directories!
Warning! Found 2,935 executable files in the PATH directories!
Warning! Found 2,935 executable files in the PATH directories!
 Debug level: Off
|-------------------------------|-----------|-----------|
|Event name                     | Time (s)  | Delta (s) |
|-------------------------------|-----------|-----------|
|start                          |   0.000   |   0.000   |
|pre_execer_init                |   0.014   |   0.014   |
|post_execer_init               |   0.058   |   0.043   |
|on_pre_rc                      |   0.058   |   0.000   |
|on_post_rc                     |   0.233   |   0.175   |
|on_ptk_create                  |   0.313   |   0.080   |
|on_post_init                   |   0.313   |   0.000   |
|on_pre_cmdloop                 |   0.313   |   0.000   |
|on_pre_prompt                  |   0.313   |   0.000   |
|on_pre_prompt_tokens           |   0.345   |   0.032   |
|on_post_prompt_tokens          |   0.356   |   0.011   |
|on_pre_prompt_pygments_lexer   |   0.356   |   0.000   |
|on_post_prompt_pygments_lexer  |   0.570   |   0.214   |
|on_pre_prompter_prompt         |   0.570   |   0.000   |
|-------------------------------|-----------|-----------|

```

</details>

After PR - 0.392:

<details>

```
(base) pc@pc:~$ COMMANDS_CACHE_SIZE_WARNING=1 xonsh --timings --rc ~/.xonshrc_conda 
Warning! Found 2,935 executable files in the PATH directories!
 Debug level: Off
|-------------------------------|-----------|-----------|
|Event name                     | Time (s)  | Delta (s) |
|-------------------------------|-----------|-----------|
|start                          |   0.000   |   0.000   |
|pre_execer_init                |   0.015   |   0.015   |
|post_execer_init               |   0.059   |   0.044   |
|on_pre_rc                      |   0.059   |   0.000   |
|on_post_rc                     |   0.217   |   0.158   |
|on_ptk_create                  |   0.301   |   0.084   |
|on_post_init                   |   0.301   |   0.000   |
|on_pre_cmdloop                 |   0.301   |   0.000   |
|on_pre_prompt                  |   0.301   |   0.000   |
|on_pre_prompt_tokens           |   0.334   |   0.033   |
|on_post_prompt_tokens          |   0.349   |   0.015   |
|on_pre_prompt_pygments_lexer   |   0.349   |   0.000   |
|on_post_prompt_pygments_lexer  |   0.392   |   0.043   |
|on_pre_prompter_prompt         |   0.392   |   0.000   |
|-------------------------------|-----------|-----------|
```

</details>

The difference is not so big like in first message example because you already merge 3913 to the master.