@scopatz 2020-10-23
Thanks @anki-code - so on Ubuntu, I am seeing this:

**before PR**
<details>
```
scopatz@artemis:~$ xonsh --timings --shell-type=ptk
 Debug level: Off
|-------------------------|-----------|-----------|
|Event name               | Time (s)  | Delta (s) |
|-------------------------|-----------|-----------|
|start                    |   0.000   |   0.000   |
|pre_execer_init          |   0.011   |   0.011   |
|post_execer_init         |   0.048   |   0.036   |
|on_pre_rc                |   0.048   |   0.000   |
|on_post_rc               |   0.178   |   0.130   |
|on_ptk_create            |   0.254   |   0.076   |
|on_post_init             |   0.254   |   0.000   |
|on_pre_cmdloop           |   0.255   |   0.000   |
|on_pre_prompt_format     |   0.255   |   0.000   |
|on_pre_prompt_tokenize   |   0.277   |   0.023   |
|on_post_prompt_tokenize  |   0.288   |   0.011   |
|on_pre_prompt_style      |   0.288   |   0.000   |
|on_post_prompt_style     |   0.417   |   0.129   |
|on_pre_prompt            |   0.417   |   0.000   |
|-------------------------|-----------|-----------|
```
</details>

**After PR**
<details>
```
scopatz@artemis:~$ xonsh --timings --shell-type=ptk
 Debug level: Off
|-------------------------|-----------|-----------|
|Event name               | Time (s)  | Delta (s) |
|-------------------------|-----------|-----------|
|start                    |   0.000   |   0.000   |
|pre_execer_init          |   0.014   |   0.014   |
|post_execer_init         |   0.050   |   0.036   |
|on_pre_rc                |   0.050   |   0.000   |
|on_post_rc               |   0.180   |   0.130   |
|on_ptk_create            |   0.259   |   0.079   |
|on_post_init             |   0.259   |   0.000   |
|on_pre_cmdloop           |   0.259   |   0.000   |
|on_pre_prompt_format     |   0.259   |   0.000   |
|on_pre_prompt_tokenize   |   0.282   |   0.023   |
|on_post_prompt_tokenize  |   0.292   |   0.010   |
|on_pre_prompt_style      |   0.292   |   0.000   |
|on_post_prompt_style     |   0.439   |   0.147   |
|on_pre_prompt            |   0.439   |   0.000   |
|-------------------------|-----------|-----------|
base scopatz@artemis ~ $
```
</details>

So this is actually 5% slower than what is on master now.