@scopatz 2020-10-24
here are my results:

Master:

<details>

```
scopatz@artemis:~$ xonsh --timings --shell-type=ptk --no-rc

                                Welcome to the xonsh shell (0.9.24.dev77)                                

                                  ~ All that is and all that shell be ~                                  

---------------------------------------------------------------------------------------------------------
xonfig tutorial    ->    Launch the tutorial in the browser
xonfig web         ->    Run the configuration tool in the browser and claim your shell 
(Note: Run the configuration tool or create a ~/.xonshrc file to suppress the welcome screen)

 Debug level: Off
|-------------------------|-----------|-----------|
|Event name               | Time (s)  | Delta (s) |
|-------------------------|-----------|-----------|
|start                    |   0.000   |   0.000   |
|pre_execer_init          |   0.016   |   0.016   |
|post_execer_init         |   0.050   |   0.035   |
|on_pre_rc                |   0.050   |   0.000   |
|on_post_rc               |   0.051   |   0.000   |
|on_ptk_create            |   0.119   |   0.068   |
|on_post_init             |   0.119   |   0.000   |
|on_pre_cmdloop           |   0.173   |   0.054   |
|on_pre_prompt_format     |   0.173   |   0.000   |
|on_pre_prompt_tokenize   |   0.173   |   0.000   |
|on_post_prompt_tokenize  |   0.270   |   0.097   |
|on_pre_prompt_style      |   0.270   |   0.000   |
|on_post_prompt_style     |   0.304   |   0.034   |
|on_pre_prompt            |   0.304   |   0.000   |
|-------------------------|-----------|-----------|
```

</details>

This PR: 

<details>

```
scopatz@artemis:~$ xonsh --timings --shell-type=ptk --no-rc

                                Welcome to the xonsh shell (0.9.24.dev86)                                

                                       ~ The shell, bourne again ~                                       

---------------------------------------------------------------------------------------------------------
xonfig tutorial    ->    Launch the tutorial in the browser
xonfig web         ->    Run the configuration tool in the browser and claim your shell 
(Note: Run the configuration tool or create a ~/.xonshrc file to suppress the welcome screen)

 Debug level: Off
|-------------------------|-----------|-----------|
|Event name               | Time (s)  | Delta (s) |
|-------------------------|-----------|-----------|
|start                    |   0.000   |   0.000   |
|pre_execer_init          |   0.011   |   0.011   |
|post_execer_init         |   0.047   |   0.036   |
|on_pre_rc                |   0.047   |   0.000   |
|on_post_rc               |   0.047   |   0.000   |
|on_ptk_create            |   0.112   |   0.065   |
|on_post_init             |   0.113   |   0.000   |
|on_pre_cmdloop           |   0.168   |   0.055   |
|on_pre_prompt_format     |   0.168   |   0.000   |
|on_pre_prompt_tokenize   |   0.168   |   0.000   |
|on_post_prompt_tokenize  |   0.263   |   0.096   |
|on_pre_prompt_style      |   0.263   |   0.000   |
|on_post_prompt_style     |   0.296   |   0.032   |
|on_pre_prompt            |   0.296   |   0.000   |
|-------------------------|-----------|-----------|
```

</details>

Huh, so that is faster with `--no-rc` by 2.6%....  It must be coming from my xonshrc then.